Change to BigDecimal successful!

+++ src/uk/ac/aber/dcs/aberpizza/data/Item.java
@@ -1,7 +1,9 @@
 package uk.ac.aber.dcs.aberpizza.data;
 
+import java.math.BigDecimal;
+
 public interface Item {
-	public Double getPrice();
+	public BigDecimal getPrice();
 	public void setPrice(Double p);
 	public String getDescription();
 	public void setDescription(String d);
